requirements.txt
flask==3.1.1
gunicorn==23.0.0
authlib==1.3.2
requests==2.32.3
python-dotenv==1.0.1
psycopg2-binary==2.9.10
